+++ RELEASES.md
@@ -633,3 +633,13 @@ Based on:
 - [typescript v5.2.0-alpha.29] .
 ### Releases
 - [NPM v5.2.0-alpha.29] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.29 - .
+
+## 2025-08-06 13:59:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.519.0 (2.552.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v5.2.0-alpha.31] .
+### Releases
+- [NPM v5.2.0-alpha.31] https://www.npmjs.com/package/meep-moop/v/5.2.0-alpha.31 - .